Expand sanitizer and fuzz coverage across the runtime trust boundaries
The sanitizer job becomes a matrix: ASan+UBSan on GCC and on Clang, and
ThreadSanitizer on Clang, each with CC set, the compiler identity and the
linked runtimes proven by scripts/ci/check_sanitizer_build.sh, the
syntax-check boot inside the instrumented suite, and the installed server
booted through autorun with a pre-authentication client conversation
(scripts/ci/smoke_client.py) before a graceful shutdown. Logs are uploaded
on failure.

A production-linked libFuzzer harness, luminari_fuzz, links the cutest
objects so each target drives the parser the server compiles: dotenv,
lib/etc/config, DG script expressions and variables, world/mobile/object/
zone/trigger records, socket line assembly and command tokenizers,
Intermud3 framing and JSON-RPC, AI provider responses, the Discord bridge
JSON, and the web onboarding controls. The loader and config targets run
behind an exit guard because those parsers end the process on a rejected
file. Seeds, dictionaries, and regression inputs are checked in; the
production-linked suite replays all of them deterministically in a forked
child. scripts/ci/run_fuzz_targets.sh runs every target, keeps and
minimizes reproducers, and the new fuzz job runs the replay, 15 seconds per
target, and the standalone protocol and binary format fuzzers, uploading
fuzz-artifacts on failure. fuzz-campaign.yml runs longer weekly campaigns.

Defects the first campaigns found, each with a saved regression input:

- asciiflag_conv_aff shifted a 32-bit int by more than its width for
  uppercase affect letters from F on;
- a mob Feat or MFeat espec indexed the feat arrays with any number;
- check_object read a NULL short description;
- parse_trigger tokenized a NULL command list for an empty script;
- flush_queues left a descriptor pointing at a pooled large output buffer,
  so a second "--" linked the pool into a cycle and hung the server;
- var_subst stepped past the terminator of an unterminated %variable and
  read beyond its stack buffer when the line filled it;
- script arithmetic overflowed int, and load_default_config leaked or
  double-freed its strings when the defaults were loaded again.

docs/guides/TESTING_GUIDE.md documents the jobs, the reproduction
commands, the supported compilers from the compiler policy, the
no-suppression rule, the MemorySanitizer assessment, and OSS-Fuzz
suitability.

+++ src/dgscript/dg_scripts.c
@@ -1870,17 +1870,19 @@ static void eval_op(const char *op, char *lhs, char *rhs, char *result,
   else if (!strcmp("/=", op))
     sprintf(result, "%c", str_str(lhs, rhs) ? '1' : '0');
 
+  /* Script operands are ints, but their product or sum need not fit one;
+   * evaluate in a wider type so a script cannot overflow the evaluator. */
   else if (!strcmp("*", op))
-    sprintf(result, "%d", atoi(lhs) * atoi(rhs));
+    sprintf(result, "%lld", (long long)atoi(lhs) * atoi(rhs));
 
   else if (!strcmp("/", op))
-    sprintf(result, "%d", (n = atoi(rhs)) ? (atoi(lhs) / n) : 0);
+    sprintf(result, "%lld", (n = atoi(rhs)) ? ((long long)atoi(lhs) / n) : 0LL);
 
   else if (!strcmp("+", op))
-    sprintf(result, "%d", atoi(lhs) + atoi(rhs));
+    sprintf(result, "%lld", (long long)atoi(lhs) + atoi(rhs));
 
   else if (!strcmp("-", op))
-    sprintf(result, "%d", atoi(lhs) - atoi(rhs));
+    sprintf(result, "%lld", (long long)atoi(lhs) - atoi(rhs));
 
   else if (!strcmp("!", op))
   {
